Partially added authorization condition

+++ DiscordClone/Controllers/GroupsController.cs
@@ -12,8 +12,8 @@ namespace DiscordClone.Controllers
     // Le decomentam la testare
 
 
-    // [Authorize]
-   
+    [Authorize]
+
 
     public class GroupsController : BaseController
     {
@@ -35,14 +35,14 @@ namespace DiscordClone.Controllers
             _env = env;
         }
         //[Authorize(Roles ="User,Moderator,Admin")]
-       
-       
+
+
 
         public IActionResult Show(int id)
         {
-                
 
-            
+
+
             var group = db.Groups.Where(o => o.Id == id).FirstOrDefault();
             if (group == null)
             {
@@ -50,41 +50,42 @@ namespace DiscordClone.Controllers
                 TempData["messageType"] = "alert-danger";
                 return RedirectToAction("Groups", "Index");
             }
+
             var t = db.Users.Where(o => o.Id == group.UserId).FirstOrDefault().UserName;
             if (t != null)
                 ViewBag.Owner = t;
 
-            
-            
+
+
             return View(group);
         }
 
 
         // Se afiseaza lista tuturor grupurilor in care nu te afli + categoria lor
         // Acum se afiseaza toate grupurile, modific filtrarea mai incolo
-        //[Authorize(Roles = "User, Moderator, Admin")]
+        [Authorize(Roles = "User, Moderator, Admin")]
         public IActionResult Index()
         {
 
             ViewBag.ToateGrupurile = db.Groups;
-            
+
             var userId = _userManager.GetUserId(User);
-            var allGroupsImin =  db.UserGroups.Where(o=>userId == o.UserId).Select(o=>o.GroupId).ToList();
-            var groups = db.Groups.Where(o=> !allGroupsImin.Contains(o.Id.ToString())).ToList();
+            var allGroupsImin = db.UserGroups.Where(o => userId == o.UserId).Select(o => o.GroupId).ToList();
+            var groups = db.Groups.Where(o => !allGroupsImin.Contains(o.Id.ToString())).ToList();
             var groupsIds = groups.Select(o => o.Id).ToList();
-           
+
             // var groups = db.Groups.Where(o => allGroupsNotIn.Contains(o.Id.ToString()));
-            
-            
-            
-            if( TempData.ContainsKey("message"))
+
+
+
+            if (TempData.ContainsKey("message"))
             {
                 ViewBag.Message = TempData["message"].ToString();
                 // ViewBag.Alert = TempData["messageType"].ToString();
             }
 
             ViewBag.Categories = GetAllCategories();
-            
+
 
             var search = "";
 
@@ -93,11 +94,12 @@ namespace DiscordClone.Controllers
                 search = Convert.ToString(HttpContext.Request.Query["search"]).Trim();
 
                 List<int> groupsId = db.Groups.Where(at => at.Name.Contains(search)).Select(a => a.Id).ToList();
-                var groups2 = db.Groups.Where(o => groupsId.Contains(o.Id) && groupsIds.Contains(o.Id)).OrderByDescending( a => a.Name).ToList();
+                var groups2 = db.Groups.Where(o => groupsId.Contains(o.Id) && groupsIds.Contains(o.Id))
+                    .OrderByDescending(a => a.Name).ToList();
                 groups = groups2;
             }
 
-            if( search != "")
+            if (search != "")
             {
                 ViewBag.PaginationBaseUrl = "/Groups/Index/?search=" + search;
 
@@ -114,11 +116,11 @@ namespace DiscordClone.Controllers
         }
 
         // Se afiseaza formularul in care vei crea un grup 
-        //[Authorize(Roles = "User, Moderator, Admin")]
+        [Authorize(Roles = "User, Moderator, Admin")]
         public IActionResult New()
         {
-           
-            
+
+
             Group group = new Group();
 
             group.Categ = ViewBag.Categories ?? GetAllCategories();
@@ -126,9 +128,9 @@ namespace DiscordClone.Controllers
             return View(group);
         }
 
-        //[Authorize(Roles ="User, Moderator, Admin")]
+        [Authorize(Roles = "User, Moderator, Admin")]
         [HttpPost]
-        public async Task<IActionResult> New( Group group, IFormFile ImageRPath)
+        public async Task<IActionResult> New(Group group, IFormFile ImageRPath)
         {
 
 
@@ -142,11 +144,12 @@ namespace DiscordClone.Controllers
             }
             else
             {
-                var allowedExtensions = new[] { ".jpg", ".jpeg", ".png"};
+                var allowedExtensions = new[] { ".jpg", ".jpeg", ".png" };
                 var fileExtension = Path.GetExtension(ImageRPath.FileName).ToLower();
                 if (!allowedExtensions.Contains(fileExtension))
                 {
-                    ModelState.AddModelError("ArticleImage", "Fișierul trebuie să fie o imagine (jpg, jpeg, png, gif) sau un video (mp4,  mov).");
+                    ModelState.AddModelError("ArticleImage",
+                        "Fișierul trebuie să fie o imagine (jpg, jpeg, png, gif) sau un video (mp4,  mov).");
                     return View(group);
                 }
 
@@ -164,21 +167,22 @@ namespace DiscordClone.Controllers
                 group.ImageRPath = databaseFileName;
 
             }
-            if ( ModelState.IsValid )
+
+            if (ModelState.IsValid)
             {
-                
-                
+
+
                 Channel channel = new Channel();
                 channel.Name = "Canalul: " + group.Name;
                 channel.Description = "Acesta este canalul default";
                 channel.UserId = _userManager.GetUserId(User);
-                
+
                 group.UserId = _userManager.GetUserId(User);
                 group.Date = DateTime.Now;
-                
+
                 db.Groups.Add(group);
                 db.SaveChanges();
-                
+
                 var userGroup = new UserGroup();
                 userGroup.GroupId = group.Id.ToString();
                 userGroup.UserId = _userManager.GetUserId(User);
@@ -186,14 +190,14 @@ namespace DiscordClone.Controllers
                 userGroup.Culoare = "#F9C74F";
                 db.UserGroups.Add(userGroup);
                 db.SaveChanges();
-                
+
                 channel.GroupId = group.Id;
                 db.Channels.Add(channel);
                 db.SaveChanges();
 
                 group.GroupChannelId = channel.Id;
                 db.SaveChanges();
-                
+
                 TempData["message"] = "S-a creat un grup";
                 TempData["messageType"] = "alert-succes";
                 return RedirectToAction("Index", "Groups");
@@ -204,36 +208,51 @@ namespace DiscordClone.Controllers
                 return View(group);
             }
         }
-
+        
+        [Authorize]
         public IActionResult Edit(int id)
         {
-           
 
+
+            var userCurrentId = _userManager.GetUserId(User);
+            var currentGrupId = db.Groups.Find(id).Id;
+            var currentUserRole = db.UserGroups.Where(o=>o.GroupId == currentGrupId.ToString() && o.UserId == userCurrentId).First().Role;
+            if(currentUserRole != "Moderator")
+                return RedirectToAction("Index", "Groups");
             
+            
+
             var group = db.Groups.Where(o => o.Id == id).FirstOrDefault();
             if (group == null)
             {
                 TempData["message"] = "Nu exista grupul pe care incerci sa-l editezi";
                 TempData["messageType"] = "alert-danger";
                 return RedirectToAction("Index", "Groups");
-                
+
             }
+
             group.Categ = GetAllCategories();
             return View(group);
-            
+
         }
 
         [HttpPost]
         public IActionResult Edit(Group editedGroup)
         {
-           
+
+            var userCurrentId = _userManager.GetUserId(User);
+            var currentGrupId = db.Groups.Find(editedGroup.Id).Id;
+            var currentUserRole = db.UserGroups.Where(o=>o.GroupId == currentGrupId.ToString() && o.UserId == userCurrentId).First().Role;
+            if(currentUserRole != "Moderator")
+                return RedirectToAction("Index", "Groups");
+            
             
             ///
             /// Mie imi sterge poza cand dau sa editeze, si nu pot adauga alta
             ///
             var group = db.Groups.Where(o => o.Id == editedGroup.Id).FirstOrDefault();
-            
-            if (group!=null && ModelState.IsValid)
+
+            if (group != null && ModelState.IsValid)
             {
                 group.Name = editedGroup.Name;
                 group.Description = editedGroup.Description;
@@ -251,7 +270,7 @@ namespace DiscordClone.Controllers
                 TempData["message"] = "Nu exista grupul pe care incerci sa-l editezi";
                 TempData["messageType"] = "alert-danger";
                 return RedirectToAction("Index", "Groups");
-                
+
             }
             else
             {
@@ -262,22 +281,27 @@ namespace DiscordClone.Controllers
 
         [HttpPost]
         public IActionResult Delete(int id)
-        {   
-            
+        {
+            var group = db.Groups.Where(o => o.Id == id).FirstOrDefault();
+            var userCurrentId = _userManager.GetUserId(User);
+            var currentGrupId = db.Groups.Find(id).Id;
+            var currentUserRole = db.UserGroups.Where(o=>o.GroupId == currentGrupId.ToString() && o.UserId == userCurrentId).First().Role;
+            if(!((currentUserRole == "Moderator"&& group.UserId == userCurrentId) || User.IsInRole("Admin")))
+                return RedirectToAction("Index", "Groups");
 
-            
             //// Aici am adaugat comentariu pt ca ar trebui sa stergem si canalele asociate grupului pe care-l facem
 
-            var group = db.Groups.Where(o => o.Id == id).FirstOrDefault();
-            var channels = db.Channels.Where( o => o.GroupId == id).FirstOrDefault();
+           
+            var channels = db.Channels.Where(o => o.GroupId == id).FirstOrDefault();
             if (group == null)
             {
                 TempData["message"] = "Grupul pe care incerci sa-l stergi nu exista";
                 TempData["messageType"] = "alert-danger";
             }
+
             db.Groups.Remove(group);
             //Am facut un check sa vad daca este null channels ca altfel da eroare
-            if(channels!=null)
+            if (channels != null)
                 db.Channels.Remove(channels);
             db.SaveChanges();
             TempData["message"] = "Grupul a fost sters cu succes";
@@ -290,7 +314,7 @@ namespace DiscordClone.Controllers
         {
             var selectList = new List<SelectListItem>();
             var categories = from cat in db.Categories
-                             select cat;
+                select cat;
 
             foreach (var category in categories)
             {
@@ -300,9 +324,10 @@ namespace DiscordClone.Controllers
 
                 selectList.Add(listItem);
             }
+
             return selectList;
         }
-
-       
-    }
+        
+        
+}
 }
